% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftConnection.R
\name{close_rs_connection}
\alias{close_rs_connection}
\title{Close a redshift connection}
\usage{
close_rs_connection(rs_conn)
}
\arguments{
\item{rs_conn}{An R Redshift connection (as given by the package RPostgres).}
}
\description{
This function allows you to close a Redshift connection.
}
\examples{
close_rs_connection(rs_conn)
}
\keyword{postgres}
\keyword{query,}
\keyword{redshift,}
\keyword{rs,}
\keyword{sql,}
