% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftConnection.R
\name{create_rs_connection}
\alias{create_rs_connection}
\title{An R Redshift connection}
\usage{
create_rs_connection(json_file = "")
}
\arguments{
\item{jsonFile}{File name of the file containing the keys. Defaults to ''.}
}
\value{
A redshift connection.
}
\description{
This function allows you to create a connection between R and Redshift given an appropriate 
keys file. Connection created with the package RPostgres.
}
\examples{
create_rs_connection('~/keys/envs.json')
}
\keyword{redshift,}
\keyword{rs}
