% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshiftConnection.R
\name{df_from_query}
\alias{df_from_query}
\title{The result of a redshift query}
\usage{
df_from_query(rs_conn, query = "SELECT COUNT(*) FROM dims.opportunities")
}
\arguments{
\item{rs_conn}{An R Redshift connection (as given by the package RPostgres).}

\item{query}{A Redshift query.}
}
\value{
A tibble that stores the query results.
}
\description{
This function allows you to create a dataframe out of the result of a given redshift query.
}
\examples{
df_from_query(rs_conn, query = "SELECT COUNT(*) FROM dims.opportunities")
}
\keyword{postgres}
\keyword{query,}
\keyword{redshift,}
\keyword{rs,}
\keyword{sql,}
